[FEAT] 도장판 상세 화면 UI 컴포넌트 분리
issue: #36

도장판 상세 화면의 각 UI 요소들을 재활용 가능한 컴포넌트로 분리하여 추출
