+++ app/src/main/java/com/polzzak_android/presentation/main/model/MissionModel.kt
@@ -1,0 +1,6 @@
+package com.polzzak_android.presentation.main.model
+
+data class MissionModel(
+    val id: Int,
+    val content: String
+)
